refactor(channel): complete core/runtime extraction - final verification
