+++ crates/app/src/channel/runtime/mod.rs
@@ -4,3 +4,4 @@ pub(in crate::channel) mod http;
 pub(in crate::channel) mod serve;
 pub(in crate::channel) mod state;
 pub(in crate::channel) mod turn_feedback;
+pub(in crate::channel) mod types;
